backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.5.0

sqlalchemy>=2.0.36
# sqlite requires no driver
alembic==1.13.0

redis==5.0.1
rq==1.15.1

pandas>=2.2.0
numpy>=1.26.0

scikit-learn>=1.4.0
joblib>=1.3.0
xgboost==2.0.3
statsmodels==0.14.1
scipy>=1.11.0

# Optional: XAI libraries (for explainable AI features)
# shap>=0.43.0  # Uncomment to enable SHAP explanations
# lime>=0.2.0.1  # Uncomment to enable LIME explanations
# tensorflow>=2.15.0  # Uncomment to enable LSTM trajectory models (optional, XGBoost works without it)

python-dotenv==1.0.0
pytz==2023.3
zipfile36==0.1.3

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Serverless deployment
mangum==0.17.0

# AI/LLM support for Zenith (free options)
groq>=0.4.0  # Free, fast AI API (recommended)
requests>=2.31.0  # For Hugging Face API alternative
cachetools==5.3.2
google-generativeai>=0.3.0 # For Gemini Vision
Pillow>=10.0.0
